requirements.txt
s3fs==2021.6.1
fsspec==2021.6.1
absl-py==0.11.0
astor==0.8.1
beautifulsoup4==4.9.3
biopython==1.78
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
#dataclasses==0.8
filelock==3.0.12
gast==0.4.0
gensim==3.8.3
google-pasta==0.2.0
grpcio==1.36.1
h5py==3.1.0
idna==2.10
importlib-metadata==3.7.2
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.6.1
numpy==1.19.5
packaging==20.9
protobuf==3.15.5
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scipy==1.5.4
six==1.15.0
smart-open==4.2.0
tensorboard==1.14.0
tensorboardX==2.2
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tokenizers==0.10.1
torch==1.8.0
tqdm==4.59.0
transformers==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
